double to float, cast

+++ genann.h
@@ -36,12 +36,12 @@ extern "C" {
 #ifndef GENANN_RANDOM
 /* We use the following for uniform random numbers between 0 and 1.
  * If you have a better function, redefine this macro. */
-#define GENANN_RANDOM() (((double)rand())/RAND_MAX)
+#define GENANN_RANDOM() (((float)rand())/RAND_MAX)
 #endif
 
 struct genann;
 
-typedef double (*genann_actfun)(const struct genann *ann, double a);
+typedef float (*genann_actfun)(const struct genann *ann, float a);
 
 typedef struct genann {
     /* How many inputs, outputs, and hidden neurons. */
@@ -60,13 +60,13 @@ typedef struct genann {
     int total_neurons;
 
     /* All weights (total_weights long). */
-    double *weight;
+	float *weight;
 
     /* Stores input array and output of each neuron (total_neurons long). */
-    double *output;
+	float *output;
 
     /* Stores delta of each hidden and output neuron (total_neurons - inputs long). */
-    double *delta;
+	float *delta;
 
 } genann;
 
@@ -86,19 +86,19 @@ genann *genann_copy(genann const *ann);
 void genann_free(genann *ann);
 
 /* Runs the feedforward algorithm to calculate the ann's output. */
-double const *genann_run(genann const *ann, double const *inputs);
+float const *genann_run(genann const *ann, float const *inputs);
 
 /* Does a single backprop update. */
-void genann_train(genann const *ann, double const *inputs, double const *desired_outputs, double learning_rate);
+void genann_train(genann const *ann, float const *inputs, float const *desired_outputs, float learning_rate);
 
 /* Saves the ann. */
 void genann_write(genann const *ann, FILE *out);
 
 void genann_init_sigmoid_lookup(const genann *ann);
-double genann_act_sigmoid(const genann *ann, double a);
-double genann_act_sigmoid_cached(const genann *ann, double a);
-double genann_act_threshold(const genann *ann, double a);
-double genann_act_linear(const genann *ann, double a);
+float genann_act_sigmoid(const genann *ann, float a);
+float genann_act_sigmoid_cached(const genann *ann, float a);
+float genann_act_threshold(const genann *ann, float a);
+float genann_act_linear(const genann *ann, float a);
 
 
 #ifdef __cplusplus
